projeto_delicias_do_dcc/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-allauth==65.3.0
djangorestframework==3.15.2
sqlparse==0.5.2
